+++ packages/editor/js/extensions/libs/flex-child/test/flex-child.e2e.cy.js
@@ -1,10 +1,9 @@
 import {
-	savePage,
-	getWPDataObject,
-	getSelectedBlock,
-	redirectToFrontPage,
-	appendBlocks,
 	createPost,
+	appendBlocks,
+	setDeviceType,
+	setInnerBlock,
+	setParentBlock,
 } from '@blockera/dev-cypress/js/helpers';
 
 describe('Flex Child', () => {
@@ -44,7 +43,56 @@ describe('Flex Child', () => {
 		});
 
 		it('should have flex-child block section', () => {
+			// Group block should not have the flex child block section
+			cy.getBlock('core/group').click();
+			cy.getByDataTest('style-tab').click();
+			cy.contains('Flex Child').should('not.exist');
+
+			// Paragraph block should have the flex child block section
+			cy.getBlock('core/paragraph').click();
+			cy.getByDataTest('style-tab').click();
 			cy.contains('Flex Child').should('exist');
+
+			// Switch to tablet device
+			setDeviceType('Mobile Portrait');
+			cy.getBlock('core/group').click();
+			cy.getParentContainer('Display').within(() => {
+				cy.getByAriaLabel('Block').click();
+			});
+			cy.getBlock('core/paragraph').click();
+			cy.getByDataTest('style-tab').click();
+			cy.contains('Flex Child').should('not.exist');
+
+			// Flex child should exist on desktop device
+			setDeviceType('Desktop');
+			cy.getBlock('core/group').click();
+			cy.getBlock('core/paragraph').click();
+			cy.getByDataTest('style-tab').click();
+			cy.contains('Flex Child').should('exist');
+		});
+
+		it('should have flex-child block section', () => {
+			cy.contains('Flex Child').should('exist');
+
+			cy.getParentContainer('Display').within(() => {
+				cy.getByAriaLabel('Flex').click();
+			});
+
+			// Check inner block to make sure it has the flex child block section
+			setInnerBlock('elements/bold');
+			cy.getByDataTest('style-tab').click();
+			cy.contains('Flex Child').should('exist');
+
+			// Switch back to the parent block
+			setParentBlock();
+			cy.getParentContainer('Display').within(() => {
+				cy.getByAriaLabel('Block').click();
+			});
+
+			// Check inner block to make sure it does not have the flex child block section
+			setInnerBlock('elements/bold');
+			cy.getByDataTest('style-tab').click();
+			cy.contains('Flex Child').should('not.exist');
 		});
 	});
 });